feat(query-core, react-query): add types, deprecate `isLoading` in favor of `isPending` in mutation observer results
